Can now switch battlers
You can now switch battler while in a battle and all data is saved. Tested the creation of a battle scene with the saving and loading system and it works. Will work on either different types of moves or making the opponent able to attack you on their turn. Will also work out namespaces soon.